+++ examples/btc_anchoring_sync.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use anyhow::{anyhow, bail};
+use async_trait::async_trait;
 use bitcoincore_rpc::{Auth as BitcoinRpcAuth, Client as BitcoinRpcClient};
 use exonum::crypto::Hash;
 use exonum_btc_anchoring::{
@@ -21,12 +23,19 @@ use exonum_btc_anchoring::{
     config::Config as AnchoringConfig,
     sync::{AnchoringChainUpdateTask, ChainUpdateError, SyncWithBitcoinError, SyncWithBitcoinTask},
 };
-use exonum_cli::io::{load_config_file, save_config_file};
 use serde::{de::DeserializeOwned, ser::Serialize};
 use serde_derive::{Deserialize, Serialize};
 use structopt::StructOpt;
+use tokio::time::delay_for;
 
-use std::{collections::HashMap, convert::TryFrom, path::PathBuf, time::Duration};
+use std::{
+    collections::HashMap,
+    convert::TryFrom,
+    fs::{self, File},
+    io::prelude::*,
+    path::{Path, PathBuf},
+    time::Duration,
+};
 
 /// Client implementation for the API of the anchoring service instance.
 #[derive(Debug, Clone)]
@@ -55,18 +64,19 @@ impl ApiClient {
         format!("{}/{}", self.prefix, name.as_ref())
     }
 
-    fn get<R>(&self, endpoint: &str) -> Result<R, String>
+    async fn get<R>(&self, endpoint: &str) -> Result<R, reqwest::Error>
     where
         R: DeserializeOwned + Send + 'static,
     {
         self.client
             .get(&self.endpoint(endpoint))
             .send()
-            .and_then(|mut request| request.json())
-            .map_err(|e| e.to_string())
+            .await?
+            .json()
+            .await
     }
 
-    fn get_query<Q, R>(&self, endpoint: &str, query: &Q) -> Result<R, String>
+    async fn get_query<Q, R>(&self, endpoint: &str, query: &Q) -> Result<R, reqwest::Error>
     where
         Q: Serialize,
         R: DeserializeOwned + Send + 'static,
@@ -75,11 +85,12 @@ impl ApiClient {
             .get(&self.endpoint(endpoint))
             .query(query)
             .send()
-            .and_then(|mut request| request.json())
-            .map_err(|e| e.to_string())
+            .await?
+            .json()
+            .await
     }
 
-    fn post<Q, R>(&self, endpoint: &str, body: &Q) -> Result<R, String>
+    async fn post<Q, R>(&self, endpoint: &str, body: &Q) -> Result<R, reqwest::Error>
     where
         Q: Serialize,
         R: DeserializeOwned + Send + 'static,
@@ -88,36 +99,41 @@ impl ApiClient {
             .post(&self.endpoint(endpoint))
             .json(&body)
             .send()
-            .and_then(|mut request| request.json())
-            .map_err(|e| e.to_string())
+            .await?
+            .json()
+            .await
     }
 }
 
+#[async_trait]
 impl PrivateApi for ApiClient {
-    type Error = String;
+    type Error = reqwest::Error;
 
-    fn sign_input(&self, sign_input: SignInput) -> Result<Hash, Self::Error> {
-        self.post("sign-input", &sign_input)
+    async fn sign_input(&self, sign_input: SignInput) -> Result<Hash, Self::Error> {
+        self.post("sign-input", &sign_input).await
     }
 
-    fn add_funds(&self, transaction: btc::Transaction) -> Result<Hash, Self::Error> {
-        self.post("add-funds", &transaction)
+    async fn add_funds(&self, transaction: btc::Transaction) -> Result<Hash, Self::Error> {
+        self.post("add-funds", &transaction).await
     }
 
-    fn anchoring_proposal(&self) -> Result<AnchoringProposalState, Self::Error> {
-        self.get("anchoring-proposal")
+    async fn anchoring_proposal(&self) -> Result<AnchoringProposalState, Self::Error> {
+        self.get("anchoring-proposal").await
     }
 
-    fn config(&self) -> Result<AnchoringConfig, Self::Error> {
-        self.get("config")
+    async fn config(&self) -> Result<AnchoringConfig, Self::Error> {
+        self.get("config").await
     }
 
-    fn transaction_with_index(&self, index: u64) -> Result<Option<btc::Transaction>, Self::Error> {
-        self.get_query("transaction", &IndexQuery { index })
+    async fn transaction_with_index(
+        &self,
+        index: u64,
+    ) -> Result<Option<btc::Transaction>, Self::Error> {
+        self.get_query("transaction", &IndexQuery { index }).await
     }
 
-    fn transactions_count(&self) -> Result<AnchoringChainLength, Self::Error> {
-        self.get("transactions-count")
+    async fn transactions_count(&self) -> Result<AnchoringChainLength, Self::Error> {
+        self.get("transactions-count").await
     }
 }
 
@@ -192,6 +208,26 @@ impl SyncConfig {
             .next()
             .map(|key| key.0.network)
     }
+
+    fn load(path: impl AsRef<Path>) -> anyhow::Result<Self> {
+        let mut file = File::open(path)?;
+        let mut toml = String::new();
+        file.read_to_string(&mut toml)?;
+        toml::de::from_str(&toml).map_err(From::from)
+    }
+
+    fn save(&self, path: impl AsRef<Path>) -> anyhow::Result<()> {
+        let path = path.as_ref();
+
+        if let Some(dir) = path.parent() {
+            fs::create_dir_all(dir)?;
+        }
+
+        let mut file = File::create(path)?;
+        let value_toml = toml::Value::try_from(&self)?;
+        file.write_all(value_toml.to_string().as_bytes())?;
+        Ok(())
+    }
 }
 
 /// `Bitcoind` rpc configuration.
@@ -218,7 +254,7 @@ impl TryFrom<BitcoinRpcConfig> for BitcoinRpcClient {
 }
 
 impl GenerateConfigCommand {
-    fn run(self) -> Result<(), failure::Error> {
+    fn run(self) -> anyhow::Result<()> {
         let bitcoin_keypair = btc::gen_keypair(self.bitcoin_network);
 
         let bitcoin_rpc_config = self.bitcoin_rpc_config();
@@ -229,7 +265,7 @@ impl GenerateConfigCommand {
             bitcoin_rpc_config,
         };
 
-        save_config_file(&sync_config, self.output)?;
+        sync_config.save(self.output)?;
         log::info!("Generated initial configuration for the btc anchoring sync util.");
         log::trace!(
             "Available Bitcoin keys in key pool: {:?}",
@@ -250,9 +286,8 @@ impl GenerateConfigCommand {
 }
 
 impl RunCommand {
-    fn run(self) -> Result<(), failure::Error> {
-        let sync_config: SyncConfig = load_config_file(self.config)?;
-        // TODO rewrite on top of tokio or runtime crate [ECR-3222]
+    async fn run(self) -> anyhow::Result<()> {
+        let sync_config = SyncConfig::load(self.config)?;
         let client = ApiClient::new(sync_config.exonum_private_api, sync_config.instance_name);
         let chain_updater =
             AnchoringChainUpdateTask::new(sync_config.bitcoin_key_pool, client.clone());
@@ -264,7 +299,7 @@ impl RunCommand {
 
         let mut latest_synced_tx_index: Option<u64> = None;
         loop {
-            match chain_updater.process() {
+            match chain_updater.process().await {
                 Ok(_) => {}
                 // Client problems most often occurs due to network problems.
                 Err(ChainUpdateError::Client(e)) => {
@@ -279,7 +314,7 @@ impl RunCommand {
                 ),
                 // For the work of anchoring you need to replenish anchoring wallet.
                 Err(ChainUpdateError::NoInitialFunds) => {
-                    let address = match chain_updater.anchoring_config() {
+                    let address = match chain_updater.anchoring_config().await {
                         Ok(config) => config.anchoring_address(),
                         Err(e) => {
                             log::error!("An error in the anchoring API client occurred. {}", e);
@@ -302,7 +337,7 @@ impl RunCommand {
             }
 
             if let Some(relay) = bitcoin_relay.as_ref() {
-                match relay.process(latest_synced_tx_index) {
+                match relay.process(latest_synced_tx_index).await {
                     Ok(index) => latest_synced_tx_index = index,
 
                     Err(SyncWithBitcoinError::Client(e)) => {
@@ -313,7 +348,7 @@ impl RunCommand {
                         log::error!("An error in the Bitcoin relay occurred. {}", e)
                     }
 
-                    Err(SyncWithBitcoinError::UnconfirmedFundingTransaction(id)) => failure::bail!(
+                    Err(SyncWithBitcoinError::UnconfirmedFundingTransaction(id)) => bail!(
                         "Funding transaction with id {} is unconfirmed by Bitcoin network. \
                          This is a serious mistake that can break anchoring process.",
                         id
@@ -323,18 +358,19 @@ impl RunCommand {
                     Err(SyncWithBitcoinError::Internal(e)) => return Err(e),
                 }
             }
+
             // Don't perform this actions too frequent to avoid DOS attack.
-            std::thread::sleep(Duration::from_secs(5));
+            delay_for(Duration::from_secs(5)).await
         }
     }
 }
 
 impl GenerateKeypairCommand {
-    fn run(self) -> Result<(), failure::Error> {
-        let mut sync_config: SyncConfig = load_config_file(&self.config)?;
+    fn run(self) -> anyhow::Result<()> {
+        let mut sync_config = SyncConfig::load(&self.config)?;
 
         let network = sync_config.bitcoin_network().ok_or_else(|| {
-            failure::format_err!(
+            anyhow!(
                 "Unable to determine Bitcoin network type from config.\
                  Perhaps pool of keys in config is empty."
             )
@@ -345,7 +381,7 @@ impl GenerateKeypairCommand {
         sync_config
             .bitcoin_key_pool
             .extend(std::iter::once(bitcoin_keypair));
-        save_config_file(&sync_config, self.config)?;
+        sync_config.save(self.config)?;
         // Print the received Bitcoin public key to use it in scripts.
         println!("{}", bitcoin_pub_key);
         Ok(())
@@ -353,18 +389,19 @@ impl GenerateKeypairCommand {
 }
 
 impl Commands {
-    fn run(self) -> Result<(), failure::Error> {
+    async fn run(self) -> anyhow::Result<()> {
         match self {
             Commands::GenerateConfig(cmd) => cmd.run(),
             Commands::GenerateKeypair(cmd) => cmd.run(),
-            Commands::Run(cmd) => cmd.run(),
+            Commands::Run(cmd) => cmd.run().await,
         }
     }
 }
 
-fn main() -> Result<(), failure::Error> {
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
     exonum::helpers::init_logger()?;
-    Commands::from_args().run()
+    Commands::from_args().run().await
 }
 
 mod flatten_keypairs {