+++ src/lib.rs
@@ -38,13 +38,15 @@
 //! Create application with anchoring service
 //!
 //! ```rust,no_run
-//! use exonum_cli::NodeBuilder;
+//! use exonum_cli::{NodeBuilder, Spec};
 //!
-//! fn main() -> Result<(), failure::Error> {
+//! #[tokio::main]
+//! async fn main() -> anyhow::Result<()> {
 //!     exonum::helpers::init_logger()?;
 //!     NodeBuilder::new()
-//!         .with_rust_service(exonum_btc_anchoring::BtcAnchoringService)
+//!         .with(Spec::new(exonum_btc_anchoring::BtcAnchoringService))
 //!         .run()
+//!         .await
 //! }
 //! ```
 //!