+++ tests/api.rs
@@ -23,33 +23,42 @@ use exonum_btc_anchoring::{
     },
 };
 use exonum_supervisor::ConfigPropose;
+use exonum_testkit::TestKitApi;
 
-fn find_transaction(
-    anchoring_testkit: &mut AnchoringTestKit,
+fn init_testkit() -> (AnchoringTestKit, TestKitApi) {
+    let mut testkit = AnchoringTestKit::default();
+    let api = testkit.inner.api();
+    (testkit, api)
+}
+
+async fn find_transaction(
+    anchoring_testkit: &AnchoringTestKit,
+    anchoring_api: &TestKitApi,
     height: Option<Height>,
 ) -> Option<btc::Transaction> {
-    let api = anchoring_testkit.inner.api();
-    let proof = api.find_transaction(height).unwrap();
-    proof
-        .validate(&anchoring_testkit.inner.consensus_config())
-        .unwrap()
-        .map(|x| x.1)
-}
+    let proof = anchoring_api
+        .client()
+        .find_transaction(height)
+        .await
+        .unwrap();
 
-fn transaction_with_index(
-    anchoring_testkit: &mut AnchoringTestKit,
-    index: u64,
-) -> Option<btc::Transaction> {
-    anchoring_testkit
+    let validator_keys = anchoring_testkit
         .inner
-        .api()
-        .transaction_with_index(index)
-        .unwrap()
+        .consensus_config()
+        .validator_keys
+        .into_iter()
+        .map(|key| key.consensus_key)
+        .collect::<Vec<_>>();
+    proof.validate(&validator_keys).unwrap().map(|(_, tx)| tx)
 }
 
-#[test]
-fn actual_address() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+async fn transaction_with_index(api: &TestKitApi, index: u64) -> Option<btc::Transaction> {
+    api.client().transaction_with_index(index).await.unwrap()
+}
+
+#[tokio::test]
+async fn actual_address() {
+    let (mut anchoring_testkit, anchoring_api) = init_testkit();
     let anchoring_interval = anchoring_testkit
         .actual_anchoring_config()
         .anchoring_interval;
@@ -66,19 +75,17 @@ fn actual_address() {
         .inner
         .create_blocks_until(Height(anchoring_interval));
 
-    let anchoring_api = anchoring_testkit.inner.api();
-
     assert_eq!(
-        anchoring_api.actual_address().unwrap(),
+        anchoring_api.client().actual_address().await.unwrap(),
         anchoring_testkit
             .actual_anchoring_config()
             .anchoring_address()
     );
 }
 
-#[test]
-fn following_address() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn following_address() {
+    let (mut anchoring_testkit, anchoring_api) = init_testkit();
     let anchoring_interval = anchoring_testkit
         .actual_anchoring_config()
         .anchoring_interval;
@@ -112,14 +119,14 @@ fn following_address() {
     anchoring_testkit.inner.create_block();
 
     assert_eq!(
-        anchoring_testkit.inner.api().following_address().unwrap(),
+        anchoring_api.client().following_address().await.unwrap(),
         Some(following_address)
     );
 }
 
-#[test]
-fn find_transaction_regular() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn find_transaction_regular() {
+    let (mut anchoring_testkit, anchoring_api) = init_testkit();
     let anchoring_interval = anchoring_testkit
         .actual_anchoring_config()
         .anchoring_interval;
@@ -142,38 +149,48 @@ fn find_transaction_regular() {
     let tx_chain = anchoring_schema.transactions_chain;
     // Find transactions by height.
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(0))).unwrap(),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(0)))
+            .await
+            .unwrap(),
         tx_chain.get(0).unwrap()
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(3))).unwrap(),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(3)))
+            .await
+            .unwrap(),
         tx_chain.get(1).unwrap()
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(4))).unwrap(),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(4)))
+            .await
+            .unwrap(),
         tx_chain.get(1).unwrap()
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(1000))).unwrap(),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(1000)))
+            .await
+            .unwrap(),
         tx_chain.get(4).unwrap()
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, None).unwrap(),
+        find_transaction(&anchoring_testkit, &anchoring_api, None)
+            .await
+            .unwrap(),
         tx_chain.last().unwrap()
     );
     // Find transactions by index.
     for i in 0..=tx_chain.len() {
         assert_eq!(
-            transaction_with_index(&mut anchoring_testkit, i),
+            transaction_with_index(&anchoring_api, i).await,
             tx_chain.get(i)
         );
     }
 }
 
 // Check come edge cases in the find_transaction api method.
-#[test]
-fn find_transaction_configuration_change() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn find_transaction_configuration_change() {
+    let (mut anchoring_testkit, anchoring_api) = init_testkit();
     let anchoring_interval = anchoring_testkit
         .actual_anchoring_config()
         .anchoring_interval;
@@ -216,15 +233,15 @@ fn find_transaction_configuration_change() {
     let snapshot = anchoring_testkit.inner.snapshot();
     let anchoring_schema = get_anchoring_schema(&snapshot);
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(0))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(0))).await,
         anchoring_schema.transactions_chain.get(1)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(1))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(1))).await,
         anchoring_schema.transactions_chain.get(1)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, None),
+        find_transaction(&anchoring_testkit, &anchoring_api, None).await,
         anchoring_schema.transactions_chain.get(1)
     );
 
@@ -242,19 +259,19 @@ fn find_transaction_configuration_change() {
     let snapshot = anchoring_testkit.inner.snapshot();
     let anchoring_schema = get_anchoring_schema(&snapshot);
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(0))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(0))).await,
         anchoring_schema.transactions_chain.get(1)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(1))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(1))).await,
         anchoring_schema.transactions_chain.get(2)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(10))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(10))).await,
         anchoring_schema.transactions_chain.get(2)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, None),
+        find_transaction(&anchoring_testkit, &anchoring_api, None).await,
         anchoring_schema.transactions_chain.get(2)
     );
 
@@ -269,45 +286,45 @@ fn find_transaction_configuration_change() {
     let snapshot = anchoring_testkit.inner.snapshot();
     let anchoring_schema = get_anchoring_schema(&snapshot);
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(0))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(0))).await,
         anchoring_schema.transactions_chain.get(1)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(1))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(1))).await,
         anchoring_schema.transactions_chain.get(2)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(5))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(5))).await,
         anchoring_schema.transactions_chain.get(2)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, Some(Height(6))),
+        find_transaction(&anchoring_testkit, &anchoring_api, Some(Height(6))).await,
         anchoring_schema.transactions_chain.get(3)
     );
     assert_eq!(
-        find_transaction(&mut anchoring_testkit, None),
+        find_transaction(&anchoring_testkit, &anchoring_api, None).await,
         anchoring_schema.transactions_chain.get(3)
     );
 }
 
-#[test]
-fn actual_config() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn actual_config() {
+    let (anchoring_testkit, anchoring_api) = init_testkit();
+
     let cfg = anchoring_testkit.actual_anchoring_config();
 
-    let api = anchoring_testkit.inner.api();
-    assert_eq!(PublicApi::config(&api).unwrap(), cfg);
-    assert_eq!(PrivateApi::config(&api).unwrap(), cfg);
+    let client = anchoring_api.client();
+    assert_eq!(PublicApi::config(client).await.unwrap(), cfg);
+    assert_eq!(PrivateApi::config(client).await.unwrap(), cfg);
 }
 
-#[test]
-fn anchoring_proposal_ok() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn anchoring_proposal_ok() {
+    let (anchoring_testkit, anchoring_api) = init_testkit();
     let proposal = anchoring_testkit.anchoring_transaction_proposal().unwrap();
 
-    let api = anchoring_testkit.inner.api();
     assert_eq!(
-        api.anchoring_proposal().unwrap(),
+        anchoring_api.client().anchoring_proposal().await.unwrap(),
         AnchoringProposalState::Available {
             transaction: proposal.0,
             inputs: proposal.1,
@@ -315,9 +332,9 @@ fn anchoring_proposal_ok() {
     );
 }
 
-#[test]
-fn anchoring_proposal_none() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn anchoring_proposal_none() {
+    let (mut anchoring_testkit, anchoring_api) = init_testkit();
 
     // Establish anchoring transactions chain.
     anchoring_testkit.inner.create_block_with_transactions(
@@ -327,24 +344,23 @@ fn anchoring_proposal_none() {
             .flatten(),
     );
 
-    let api = anchoring_testkit.inner.api();
     assert_eq!(
-        api.anchoring_proposal().unwrap(),
+        anchoring_api.client().anchoring_proposal().await.unwrap(),
         AnchoringProposalState::None
     );
 }
 
-#[test]
-fn anchoring_proposal_err_without_initial_funds() {
+#[tokio::test]
+async fn anchoring_proposal_err_without_initial_funds() {
     let mut anchoring_testkit = AnchoringTestKit::new(4, 5);
 
     let api = anchoring_testkit.inner.api();
-    let state = api.anchoring_proposal().unwrap();
+    let state = api.client().anchoring_proposal().await.unwrap();
     assert_eq!(state, AnchoringProposalState::NoInitialFunds);
 }
 
-#[test]
-fn anchoring_proposal_err_insufficient_funds() {
+#[tokio::test]
+async fn anchoring_proposal_err_insufficient_funds() {
     let mut anchoring_testkit = AnchoringTestKit::new(4, 5);
 
     // Add an initial funding transaction to enable anchoring.
@@ -353,7 +369,7 @@ fn anchoring_proposal_err_insufficient_funds() {
         .create_block_with_transactions(anchoring_testkit.create_funding_confirmation_txs(20).0);
 
     let api = anchoring_testkit.inner.api();
-    let state = api.anchoring_proposal().unwrap();
+    let state = api.client().anchoring_proposal().await.unwrap();
     assert_eq!(
         state,
         AnchoringProposalState::InsufficientFunds {
@@ -363,9 +379,9 @@ fn anchoring_proposal_err_insufficient_funds() {
     );
 }
 
-#[test]
-fn sign_input() {
-    let mut anchoring_testkit = AnchoringTestKit::default();
+#[tokio::test]
+async fn sign_input() {
+    let (mut anchoring_testkit, anchoring_api) = init_testkit();
 
     let config = anchoring_testkit.actual_anchoring_config();
     let bitcoin_public_key = config
@@ -396,14 +412,14 @@ fn sign_input() {
         )
         .unwrap();
 
-    let tx_hash = anchoring_testkit
-        .inner
-        .api()
+    let tx_hash = anchoring_api
+        .client()
         .sign_input(SignInput {
             input: 0,
             input_signature: signature.into(),
             txid: proposal.id(),
         })
+        .await
         .unwrap();
 
     anchoring_testkit
@@ -413,18 +429,19 @@ fn sign_input() {
         .expect("Transaction should be successful");
 }
 
-#[test]
-fn add_funds_ok() {
+#[tokio::test]
+async fn add_funds_ok() {
     let anchoring_interval = 5;
     let mut anchoring_testkit = AnchoringTestKit::new(1, anchoring_interval);
+    let anchoring_api = anchoring_testkit.inner.api();
 
     let config = anchoring_testkit.actual_anchoring_config();
     let funding_transaction = create_fake_funding_transaction(&config.anchoring_address(), 10_000);
 
-    let tx_hash = anchoring_testkit
-        .inner
-        .api()
+    let tx_hash = anchoring_api
+        .client()
         .add_funds(funding_transaction)
+        .await
         .unwrap();
 
     anchoring_testkit
@@ -434,8 +451,8 @@ fn add_funds_ok() {
         .expect("Transaction should be successful");
 }
 
-#[test]
-fn add_funds_err_already_used() {
+#[tokio::test]
+async fn add_funds_err_already_used() {
     let anchoring_interval = 5;
     let mut anchoring_testkit = AnchoringTestKit::new(1, anchoring_interval);
 
@@ -454,22 +471,25 @@ fn add_funds_err_already_used() {
     anchoring_testkit
         .inner
         .api()
+        .client()
         .add_funds(funding_transaction)
+        .await
         .expect_err("Add funds must fail");
 }
 
-#[test]
-fn add_funds_err_unsuitable() {
+#[tokio::test]
+async fn add_funds_err_unsuitable() {
     let anchoring_interval = 5;
     let mut anchoring_testkit = AnchoringTestKit::new(4, anchoring_interval);
+    let anchoring_api = anchoring_testkit.inner.api();
 
     let mut config = anchoring_testkit.actual_anchoring_config();
     config.anchoring_keys.swap(1, 3);
     let funding_transaction = create_fake_funding_transaction(&config.anchoring_address(), 10_000);
 
-    anchoring_testkit
-        .inner
-        .api()
+    anchoring_api
+        .client()
         .add_funds(funding_transaction)
+        .await
         .expect_err("Add funds must fail");
 }